Shift more responsibilities to SnoopyModel
